+++ client/src/components/User/UsersList.jsx
@@ -14,8 +14,6 @@ export const UsersList = () => {
 	const [usersGroups, setUsersGroups] = useState([]);
 	const [groups, setGroups] = useState([]);
 	const [selectedUser, setSelectedUser] = useState('');
-	const [userAddGroup, setUserAddGroup] = useState('');
-	const [userRemoveGroup, setUserRemoveGroup] = useState('');
 	const [findUser, setFindUser] = useState('');
 	const { token } = useContext(AuthContext);
 	const [form, setForm] = useState({
@@ -116,40 +114,6 @@ export const UsersList = () => {
 
 	// Продумать оптимизацию filteredUser()
 
-	const addUserToGroupHandler = (event) => {
-		setUserAddGroup(event.target.value);
-	};
-
-	const removeUserFromGroup = (event) => {
-		setUserRemoveGroup(event.target.value);
-	};
-
-	const updateUserGroups = useCallback(() => {
-		if (user && usersGroups) {
-			const user_groups = usersGroups.filter(
-				(group) => group.user_id === user._id,
-			);
-
-			const userGroupsObj = user_groups[0];
-
-			const checkGroup = (groupId) => {
-				if (userGroupsObj.group_ids.length) {
-					return userGroupsObj.group_ids.indexOf(groupId);
-				}
-			};
-
-			if (checkGroup(userAddGroup) < 0 && userAddGroup)
-				userGroupsObj.group_ids.push(userAddGroup);
-
-			if (checkGroup(userRemoveGroup) >= 0 && userRemoveGroup)
-				userGroupsObj.group_ids.splice(checkGroup(userRemoveGroup), 1);
-		}
-	}, [user, usersGroups, userAddGroup, userRemoveGroup]);
-
-	useEffect(() => {
-		updateUserGroups();
-	}, [updateUserGroups]);
-
 	if (loading) {
 		return <Loader />;
 	}
@@ -163,70 +127,6 @@ export const UsersList = () => {
 						selectUserHandler={selectUserHandler}
 						getUserGroups={getUserGroups}
 					/>
-				</div>
-			</div>
-
-			<div className="widget__wrapper widget__user--update has-shadow">
-				<div className="widget__header" id={selectedUser}>
-					<Form.Group controlId="findByEmeil" className="form__find-user">
-						<Form.Label>Найти пользователя по Email</Form.Label>
-						<Form.Control
-							type="text"
-							name="find_by_emeil"
-							value={findUser}
-							onChange={(e) => setFindUser(e.target.value)}
-							onKeyPress={findUserHandler}
-						/>
-					</Form.Group>
-					<h4 className="widget__title">Редактировать пользователя</h4>
-				</div>
-				<div className="widget__body">
-					{user && <UserItem user={user} getUserGroups={getUserGroups} />}
-					<Form className="form__createGroup">
-						<Form.Group controlId="inputEditGroup" className="mb-3">
-							<Form.Label>Добавить пользователя в группу</Form.Label>
-							<Form.Control
-								as="select"
-								name="user_groups"
-								value={userAddGroup}
-								onChange={addUserToGroupHandler}
-							>
-								<option value="">Выберите группу</option>
-								{groups.map((el, i) => {
-									return (
-										<option key={i} value={el._id}>
-											{el.name}
-										</option>
-									);
-								})}
-							</Form.Control>
-						</Form.Group>
-						<Form.Group controlId="inputGroup" className="mb-3">
-							<Form.Label>Удалить пользователя из группы</Form.Label>
-							<Form.Control
-								as="select"
-								name="user_groups"
-								onChange={removeUserFromGroup}
-							>
-								<option value="">Выберите группу</option>
-								{groups.map((el, i) => {
-									return (
-										<option key={i} value={el._id}>
-											{el.name}
-										</option>
-									);
-								})}
-							</Form.Control>
-						</Form.Group>
-						<Button
-							className="btn btn-primary btn__gradient btn__grad-danger btn__sign-in"
-							type="submit"
-							disabled={loading}
-							onClick={updateUserGroups}
-						>
-							Сохранить изменения
-						</Button>
-					</Form>
 				</div>
 			</div>
 		</>
