+++ client/src/routes.js
@@ -6,6 +6,7 @@ import { GroupsPage } from './pages/GroupsPage';
 import { GroupCreatePage } from './pages/GroupCreatePage';
 import { UsersPage } from './pages/UsersPage';
 import { UserCreatePage } from './pages/UserCreatePage';
+import { UserEditPage } from './pages/UserEditPage';
 import { TrainingsPage } from './pages/TrainingsPage';
 import { TrainingCreatePage } from './pages/TrainingCreatePage';
 import { TrainingDetailPage } from './pages/TrainingDetailPage';
@@ -32,6 +33,9 @@ export const useRoutes = (isAuthenticated) => {
 				</Route>
 				<Route path="/user/create" exact>
 					<UserCreatePage />
+				</Route>
+				<Route path="/user/edit/:id">
+					<UserEditPage />
 				</Route>
 				<Route path="/trainings" exact>
 					<TrainingsPage />
